Add OTel Capabilities
(cherry picked from commit 68203cfdd2157e23978300d6067b6983fc178f71)
