+++ extensions/opentelemetry/deployment/src/main/java/io/quarkus/opentelemetry/deployment/logging/LogHandlerProcessor.java
@@ -1,8 +1,5 @@
 package io.quarkus.opentelemetry.deployment.logging;
 
-import java.util.function.BooleanSupplier;
-import java.util.function.Function;
-
 import org.jboss.jandex.DotName;
 
 import io.opentelemetry.sdk.autoconfigure.spi.logs.ConfigurableLogRecordExporterProvider;
@@ -19,12 +16,11 @@ import io.quarkus.deployment.annotations.ExecutionTime;
 import io.quarkus.deployment.annotations.Record;
 import io.quarkus.deployment.builditem.LogHandlerBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.ServiceProviderBuildItem;
-import io.quarkus.opentelemetry.runtime.config.build.OTelBuildConfig;
 import io.quarkus.opentelemetry.runtime.config.runtime.OTelRuntimeConfig;
 import io.quarkus.opentelemetry.runtime.logs.OpenTelemetryLogRecorder;
 import io.quarkus.opentelemetry.runtime.logs.spi.LogsExporterCDIProvider;
 
-@BuildSteps(onlyIf = LogHandlerProcessor.LogsEnabled.class)
+@BuildSteps(onlyIf = LogsEnabled.class)
 class LogHandlerProcessor {
 
     private static final DotName LOG_RECORD_EXPORTER = DotName.createSimple(LogRecordExporter.class.getName());
@@ -50,20 +46,5 @@ class LogHandlerProcessor {
             OTelRuntimeConfig config,
             BeanContainerBuildItem beanContainerBuildItem) {
         return new LogHandlerBuildItem(recorder.initializeHandler(beanContainerBuildItem.getValue(), config));
-    }
-
-    public static class LogsEnabled implements BooleanSupplier {
-        OTelBuildConfig otelBuildConfig;
-
-        public boolean getAsBoolean() {
-            return otelBuildConfig.logs().enabled()
-                    .map(new Function<Boolean, Boolean>() {
-                        @Override
-                        public Boolean apply(Boolean enabled) {
-                            return otelBuildConfig.enabled() && enabled;
-                        }
-                    })
-                    .orElseGet(() -> otelBuildConfig.enabled());
-        }
     }
 }
